Adverbs

Adverbs are functions that modify the behaviour of a function.

TidierIteration.compose — Function

compose(args...)

Create the composite function of the args....

compose(f1, f2)(x) is equal to f2(f1(x)).

Missing docstring.

Missing docstring for compose_n. Check Documenter's build log for details.

TidierIteration.negate — Function

negate(f)

Create the negation of f.

TidierIteration.possibly — Function

possibly(f, otherwise = nothing)

Create a modified version of the function f such that it returns otherwise when f(x) is an error.