Maps

One variable

Applying one-variable functions to collections

TidierMap.map_arr — Function

map_arr(x, f)

Apply function f to the list x.

TidierMap.map_dbl — Function

map_dbl(x, f)

Apply function f to the list x and convert the result to float.

TidierMap.map_int — Function

map_int(x, f)

Apply function f to the list x and convert the result to integer.

TidierMap.map_lgl — Function

map_lgl(x, f)

Apply function f to the list x and convert the result to logical (boolean).

TidierMap.map_chr — Function

map_chr(x, f)

Apply function f to the list x and convert the result to string (character).

TidierMap.walk — Function

walk(x, f)

Apply function f to the list x and return nothing.

Two variables

Applying two-variable functions to collections.

TidierMap.map2_arr — Function

map2_arr(x, y, f)

Apply the two-variable function f to the objects x and y.

TidierMap.map2_dbl — Function

map2_dbl(x, y, f)

Apply the two-variable function f to the objects x and y and convert it to float.

TidierMap.map2_int — Function

map2_int(x, y, f)

Apply the two-variable function f to the objects x and y and convert it to integer.

TidierMap.map2_lgl — Function

map2_lgl(x, y, f)

Apply the two-variable function f to the objects x and y and convert it to logical (boolean).

TidierMap.map2_chr — Function

map2_chr(x, y, f)

Apply the two-variable function f to the objects x and y and convert it to character (string).

TidierMap.walk2 — Function

walk2(x, y, f)

Apply the two-variable function f to the objects x and y and return nothing.

One variable and index

Applying functions to a collection and the corresponding index of the element

TidierMap.imap_arr — Function

imap_arr(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i.

TidierMap.imap_dbl — Function

imap_dbl(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it to float (double).

TidierMap.imap_int — Function

imap_int(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it to integer.

TidierMap.imap_lgl — Function

imap_lgl(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it to boolean (logical).

TidierMap.imap_chr — Function

imap_chr(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it to string (character).

TidierMap.iwalk — Function

imap_chr(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and return nothing.