Modify

Functions to modify a collection.

TidierIteration.modify — Function

modify(x, f)

Modify a copy of x applying f to each of its elements.

TidierIteration.modify! — Function

modify!(x, f)

Modify x applying f to each of its elements.

TidierIteration.modify_if — Function

modify_if(x, p, f)

Modify a copy of x applying f to each of its elements where the function p is true.

TidierIteration.modify_if! — Function

modify_if!(x, p, f)

Modify x applying f to each of its elements where the function p is true.

TidierIteration.keep — Function

keep(x, p)

Keep the elements x_i of collection x such that p(x_i) is true.

TidierIteration.keep! — Function

keep!(x, p)

Keep the elements x_i of collection x such that p(x_i) is true.

TidierIteration.discard — Function

discard(x, p)

Discard the elements x_i of collection x such that p(x_i) is true.

TidierIteration.discard! — Function

discard!(x, p)

Discard the elements x_i of collection x such that p(x_i) is true.

TidierIteration.compact — Function

compact(x)

Keep the elements x_i of collection x such that is_non_empty(x_i) is true.

TidierIteration.compact! — Function

compact!(x)

Keep the elements x_i of collection x such that is_non_empty(x_i) is true.