Maps

One variable

Applying one-variable functions to collections

TidierIteration.map_asis — Function

map_arr(x, f)

Apply function f to the list x and do no conversion.

TidierIteration.map_type — Function

map_type(x, f, T)

Apply function f to the list x and convert each element with function T.

TidierIteration.map_float — Function

map_float(x, f)(x, f)

Apply function f to the list x and convert the result to float.

TidierIteration.map_int — Function

map_int(x, f)

Apply function f to the list x and convert the result to integer.

TidierIteration.map_bool — Function

map_bool(x, f)

Apply function f to the list x and convert the result to logical (boolean).

TidierIteration.map_string — Function

map_string(x, f)

Apply function f to the list x and convert the result to string (character).

TidierIteration.walk — Function

walk(x, f)

Apply function f to the list x and return nothing.

Two variables

Applying two-variable functions to collections.

TidierIteration.map2_asis — Function

map2_asis(x, y, f)

Apply the two-variable function f to the objects x and y.

TidierIteration.map2_type — Function

map2_type(x, y, f, T)

Apply the two-variable function f to the objects x and y and convert each element using function T.

TidierIteration.map2_float — Function

map2_dbl(x, y, f)

Apply the two-variable function f to the objects x and y and convert it to float.

TidierIteration.map2_int — Function

map2_int(x, y, f)

Apply the two-variable function f to the objects x and y and convert it to integer.

TidierIteration.map2_bool — Function

map2_lgl(x, y, f)

Apply the two-variable function f to the objects x and y and convert it to boolean.

TidierIteration.map2_string — Function

map2_chr(x, y, f)

Apply the two-variable function f to the objects x and y and convert it to string.

TidierIteration.walk2 — Function

walk2(x, y, f)

Apply the two-variable function f to the objects x and y and return nothing.

One variable and index

Applying functions to a collection and the corresponding index of the element

TidierIteration.imap_asis — Function

imap_arr(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i.

TidierIteration.imap_type — Function

imap_type(x, f, T)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it using function T.

TidierIteration.imap_float — Function

imap_float(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it to float.

TidierIteration.imap_int — Function

imap_int(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it to integer.

TidierIteration.imap_bool — Function

imap_bool(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it to boolean.

TidierIteration.imap_string — Function

imap_string(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and convert it to string.

TidierIteration.iwalk — Function

iwalk(x, f)

Apply f to each pair (i, x_i) where x[i] = x_i and return nothing.